% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arima}
\alias{arima}
\title{There are modeling functions in this section
ARIMA Model}
\usage{
arima(train_data, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}
}
\value{
Get the ARIMA based model
}
\description{
There are modeling functions in this section
ARIMA Model
}
\keyword{internal}
