% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arima_boost}
\alias{arima_boost}
\title{ARIMA Boost}
\usage{
arima_boost(
  train_data,
  frequency,
  parallel,
  horizon,
  tscv_initial,
  date_rm_regex,
  back_test_spacing,
  fiscal_year_start
)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}

\item{parallel}{Parallel Version or not}

\item{horizon}{Horizon of model}

\item{tscv_initial}{tscv initialization}

\item{date_rm_regex}{Date removal Regex}

\item{back_test_spacing}{Back Testing Spacing}

\item{fiscal_year_start}{Fiscal Year Start}
}
\value{
Get the ARIMA based model
}
\description{
ARIMA Boost
}
