% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_fast_model.R
\name{arima_fast_model}
\alias{arima_fast_model}
\title{Adaptive fast ARIMA model specification}
\usage{
arima_fast_model(
  mode = "regression",
  forecast_horizon = NULL,
  frequency = NULL
)
}
\arguments{
\item{mode}{A single character string for the model mode.}

\item{forecast_horizon}{Number of periods to forecast.}

\item{frequency}{Frequency of the input series.}
}
\value{
An adaptive fast ARIMA model specification.
}
\description{
Adaptive fast ARIMA model specification
}
\keyword{internal}
