% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_fast_model.R
\name{arima_fast_model_fit_impl}
\alias{arima_fast_model_fit_impl}
\title{Fit an adaptive fast ARIMA model}
\usage{
arima_fast_model_fit_impl(x, y, forecast_horizon = NULL, frequency = NULL)
}
\arguments{
\item{x}{A data frame containing a \code{Date} column.}

\item{y}{Numeric training values.}

\item{forecast_horizon}{Number of periods to forecast.}

\item{frequency}{Frequency of the input series.}
}
\value{
A fitted adaptive fast ARIMA object.
}
\description{
Fit an adaptive fast ARIMA model
}
\keyword{internal}
