% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_fast_model.R
\name{arima_fast_model_predict_impl}
\alias{arima_fast_model_predict_impl}
\title{Predict an adaptive fast ARIMA model}
\usage{
arima_fast_model_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{A fitted adaptive fast ARIMA object.}

\item{new_data}{Future data containing a \code{Date} column.}

\item{...}{Additional arguments.}
}
\value{
Numeric predictions.
}
\description{
Predict an adaptive fast ARIMA model
}
\keyword{internal}
