% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_average_models.R
\name{average_models}
\alias{average_models}
\title{Average Models}
\usage{
average_models(
  refit_predictions_tbl,
  ensemble_predictions_tbl = NULL,
  model_train_test_tbl,
  max_model_average = 3,
  parallel_processing = NULL,
  num_cores = NULL
)
}
\arguments{
\item{refit_predictions_tbl}{individual model predictions}

\item{ensemble_predictions_tbl}{ensemble model predictions}

\item{model_train_test_tbl}{train test splits}

\item{max_model_average}{max number of models to average}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}
}
\value{
list with average predictions and fitted models
}
\description{
Average Models
}
\keyword{internal}
