% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos2_model.R
\name{chronos2_model}
\alias{chronos2_model}
\title{Chronos 2 model specification}
\usage{
chronos2_model(mode = "regression", forecast_horizon = NULL, frequency = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency of data}
}
\value{
Get Chronos 2 model
}
\description{
Chronos 2 model specification
}
\keyword{internal}
