% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos2_model.R
\name{chronos2_model_fit_impl}
\alias{chronos2_model_fit_impl}
\title{Fit Chronos 2 model}
\usage{
chronos2_model_fit_impl(x, y, forecast_horizon = NULL, frequency = NULL)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency of data}
}
\value{
Fitted Chronos 2 model object
}
\description{
Since Chronos 2 is a foundation model with no local training,
the fit step simply stores the training data and parameters
for use during prediction.
}
\keyword{internal}
