% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos_bolt_base_model.R
\name{chronos_bolt_base_model_predict_impl}
\alias{chronos_bolt_base_model_predict_impl}
\title{Bridge prediction function for Chronos Bolt Base Models}
\usage{
chronos_bolt_base_model_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}

\item{...}{Additional arguments}
}
\value{
numeric vector of predictions
}
\description{
Prepares the training and future data, calls the Chronos API via
\code{chronos_forecast()}, and returns the point predictions.
}
\keyword{internal}
