% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos_bolt_tiny_model.R
\name{chronos_bolt_tiny_model}
\alias{chronos_bolt_tiny_model}
\title{Chronos Bolt Tiny model specification}
\usage{
chronos_bolt_tiny_model(
  mode = "regression",
  forecast_horizon = NULL,
  frequency = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency of data}
}
\value{
Get Chronos Bolt Tiny model
}
\description{
Chronos Bolt Tiny model specification
}
\keyword{internal}
