% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{construct_forecast_models}
\alias{construct_forecast_models}
\title{Get a Forecast Model Functions}
\usage{
construct_forecast_models(
  full_data_tbl,
  external_regressors,
  xregs_future_values_list,
  fourier_periods,
  combo_variables,
  lag_periods,
  rolling_window_periods,
  hist_end_date,
  date_type,
  forecast_horizon,
  run_model_parallel,
  parallel_processing,
  run_deep_learning,
  frequency_number,
  models_to_run,
  models_not_to_run,
  run_ensemble_models,
  hist_periods_80,
  back_test_spacing,
  back_test_scenarios,
  date_regex,
  fiscal_year_start,
  seasonal_periods
)
}
\arguments{
\item{full_data_tbl}{Full Data Table}

\item{external_regressors}{External Regressors}

\item{xregs_future_values_list}{External Regressors Future Values}

\item{fourier_periods}{Fourier Periods}

\item{combo_variables}{Combo Variables}

\item{lag_periods}{Lag Periods}

\item{rolling_window_periods}{Rolling window periods}

\item{hist_end_date}{Historical End Date}

\item{date_type}{Date Type}

\item{forecast_horizon}{Forecast Horizon}

\item{run_model_parallel}{Run Model in Parallel}

\item{parallel_processing}{Which parallel processing to use}

\item{run_deep_learning}{Run Deep Learning model}

\item{frequency_number}{Frequency Number}

\item{models_to_run}{Models to Run}

\item{models_not_to_run}{Models not to run}

\item{hist_periods_80}{Hist Periods 80}

\item{back_test_spacing}{Back Test Spacing}

\item{back_test_scenarios}{Back Test Scenarios}

\item{date_regex}{Date Regex}

\item{fiscal_year_start}{Fiscal Year Start}

\item{seasonal_periods}{Seasonal Periods
We need to move these out using a scheduler}
}
\value{
a forecast_models function
}
\description{
Get a Forecast Model Functions
}
