% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{croston}
\alias{croston}
\title{Croston Model}
\usage{
croston(train_data, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}
}
\value{
Get the Croston based model
}
\description{
Croston Model
}
\keyword{internal}
