% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cubist}
\alias{cubist}
\title{Cubist Function}
\usage{
cubist(
  train_data,
  parallel,
  model_type = "single",
  horizon,
  tscv_initial,
  date_rm_regex,
  back_test_spacing,
  fiscal_year_start
)
}
\arguments{
\item{train_data}{Training Data}

\item{parallel}{Parallel Version or not}

\item{model_type}{"single" "ensemble" etc.}

\item{horizon}{Horizon of model}

\item{tscv_initial}{tscv initialization}

\item{date_rm_regex}{Date removal Regex}

\item{back_test_spacing}{Back Testing Spacing}

\item{fiscal_year_start}{Fiscal Year Start}
}
\value{
Get the cubist
}
\description{
Cubist Function
}
