% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_cubist.R
\name{cubist_multistep_fit_impl}
\alias{cubist_multistep_fit_impl}
\title{Bridge CUBIST Multistep Modeling function}
\usage{
cubist_multistep_fit_impl(
  x,
  y,
  committees = 1,
  neighbors = 0,
  max_rules = 10,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{committees}{committees}

\item{neighbors}{neighbors}

\item{max_rules}{max rules}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\description{
Bridge CUBIST Multistep Modeling function
}
\keyword{internal}
