% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{deepar}
\alias{deepar}
\title{Deep AR Model}
\usage{
deepar(train_data, horizon, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon of forecast}

\item{frequency}{Frequency of Data}
}
\value{
Get the DeepAR model
}
\description{
Deep AR Model
}
\keyword{internal}
