% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ets}
\alias{ets}
\title{ETS Model}
\usage{
ets(train_data, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}
}
\value{
Get the ETS model
}
\description{
ETS Model
}
