% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_batch_parallel.R
\name{exit_azure_parallel_within}
\alias{exit_azure_parallel_within}
\title{Call back function within forecast for exit}
\usage{
exit_azure_parallel_within(cl)
}
\arguments{
\item{cl}{Cluster variable from init

Sets the right configuration during exit}
}
\description{
Call back function within forecast for exit
}
