% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_model.R
\name{finetune_depth}
\alias{finetune_depth}
\title{Finetune depth parameter definition}
\usage{
finetune_depth(range = c(1L, 5L))
}
\arguments{
\item{range}{Integer vector of length 2 giving the minimum and maximum
allowable values (default: \code{c(1L, 5L)}).}
}
\value{
A \code{dials} parameter object describing the finetune depth hyperparameter.
}
\description{
Defines the tunable parameter used to control the fine-tuning depth
(number of layers) for the TimeGPT model. Like \code{finetune_steps()}, this can
be passed through \code{tune::tune()} and searched over an integer range.
}
