% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_model.R
\name{finetune_steps}
\alias{finetune_steps}
\title{Finetune steps parameter definition}
\usage{
finetune_steps(range = c(0L, 200L))
}
\arguments{
\item{range}{Integer vector of length 2 giving the minimum and maximum
allowable values (default: \code{c(0L, 200L)}).}
}
\value{
A \code{dials} parameter object describing the finetune steps hyperparameter.
}
\description{
Defines the tunable parameter used to control the number of fine-tuning steps
for the TimeGPT model. This parameter is exposed via \code{tune::tune()} and can
be searched over a specified integer range.
}
\keyword{internal}
