% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_back_test_scenario_hist_periods}
\alias{get_back_test_scenario_hist_periods}
\title{Gets the back testing scenarios}
\usage{
get_back_test_scenario_hist_periods(
  full_data_tbl,
  hist_end_date,
  back_test_scenarios,
  forecast_horizon,
  back_test_spacing
)
}
\arguments{
\item{full_data_tbl}{full data table}

\item{hist_end_date}{historical end date}

\item{back_test_scenarios}{back test scenarios}

\item{forecast_horizon}{forecast horizon}

\item{back_test_spacing}{back test spacing}
}
\value{
Returns back_test_scenarios and hist_periods_80
}
\description{
Gets back testing scenarios accounting for splits
}
