% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_back_test_spacing}
\alias{get_back_test_spacing}
\title{Gets the back testing spacing}
\usage{
get_back_test_spacing(back_test_spacing, date_type)
}
\arguments{
\item{back_test_spacing}{back_test_spacing override}

\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns back_test_spacing
}
\description{
Checks if back_test_spacing is set to auto and gets the right one
}
