% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data.R
\name{get_data_tbl}
\alias{get_data_tbl}
\title{Gets the tibble version of data}
\usage{
get_data_tbl(input_data, combo_variables, target_variable)
}
\arguments{
\item{input_data}{Original input data to function}

\item{combo_variables}{List of Combo variables}

\item{target_variable}{Variable to forecast}
}
\value{
Returns Standardized data_tbl (a data table a.k.a. tibble)
}
\description{
Converts input data into tibble
}
