% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_data.R
\name{get_data_tbl_final}
\alias{get_data_tbl_final}
\title{Function to create the data_tbl_final}
\usage{
get_data_tbl_final(
  data_tbl,
  combo_variables,
  forecast_approach,
  frequency_number
)
}
\arguments{
\item{data_tbl}{standardized data table}

\item{combo_variables}{list of combo variables}

\item{forecast_approach}{forecasting approach}

\item{frequency_number}{frequency number}
}
\value{
data_tbl_final
}
\description{
Function to create the data_tbl_final
}
