% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_date_regex}
\alias{get_date_regex}
\title{Gets the date regex for pasing}
\usage{
get_date_regex(date_type)
}
\arguments{
\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns date_regex
}
\description{
Checks if date_regex  is set if not gets right one
}
