% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_export_packages}
\alias{get_export_packages}
\title{Fetches a list of packages to be exported}
\usage{
get_export_packages()
}
\value{
Simple list of packages
}
\description{
Code gets used in multiple places where we need the packages list
}
