% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_data.R
\name{get_external_regressors}
\alias{get_external_regressors}
\title{Function to create the external_regressors}
\usage{
get_external_regressors(external_regressors, forecast_approach)
}
\arguments{
\item{external_regressors}{existing external_regressors}

\item{forecast_approach}{forecasting approach}
}
\value{
external_regressors
}
\description{
Function to create the external_regressors
}
