% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_parallel.R
\name{get_fcast_parallel}
\alias{get_fcast_parallel}
\title{Call parallel forecast model}
\usage{
get_fcast_parallel(combo_list, call_back_fn)
}
\arguments{
\item{combo_list}{Combo List}

\item{call_back_fn}{Call Back Function}
}
\value{
Forecast Object
}
\description{
Local machine parallelization of forecast model
}
