% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_batch_parallel.R
\name{get_fcast_parallel_azure}
\alias{get_fcast_parallel_azure}
\title{Call parallel forecast model for azure}
\usage{
get_fcast_parallel_azure(
  combo_list,
  call_back_fn,
  azure_batch_credentials,
  azure_batch_clusterConfig,
  run_name
)
}
\arguments{
\item{combo_list}{Combo List}

\item{call_back_fn}{Call Back Function}

\item{azure_batch_credentials}{Azure Batch Credentials}

\item{azure_batch_clusterConfig}{Azure Batch Cluster Config}

\item{run_name}{Run Name}
}
\value{
Forecast Object
}
\description{
Azure Batch parallelization of forecast model
}
