% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_fit_wkflw_best}
\alias{get_fit_wkflw_best}
\title{Gets a simple fit}
\usage{
get_fit_wkflw_best(train_data, tune_results, wflw_spec_tune)
}
\arguments{
\item{train_data}{Training Data}

\item{tune_results}{Tune results}

\item{wflw_spec_tune}{Worflow Spec after tuning}
}
\value{
simple recipie
}
\description{
Gets a simple fit
}
