% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_fit_wkflw_nocombo}
\alias{get_fit_wkflw_nocombo}
\title{Gets a simple fit}
\usage{
get_fit_wkflw_nocombo(train_data, model_spec, recipie_spec)
}
\arguments{
\item{train_data}{Training Data}

\item{model_spec}{Model Spec}

\item{recipie_spec}{Recipie Spec}
}
\value{
simple recipie
}
\description{
Gets a simple fit
}
