% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output_data.R
\name{get_forecast_negative_adjusted}
\alias{get_forecast_negative_adjusted}
\title{Get non-negative forecast}
\usage{
get_forecast_negative_adjusted(fcst, negative_fcst)
}
\arguments{
\item{fcst}{Input Data Frame}

\item{negative_fcst}{is negative forecast allowed}
}
\description{
Function to rectify forecast non-negativity
}
