% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_fourier_periods}
\alias{get_fourier_periods}
\title{Gets the right fourier periods}
\usage{
get_fourier_periods(fourier_periods, date_type)
}
\arguments{
\item{fourier_periods}{fourier_periods override}

\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns fourier_periods
}
\description{
Checks if fourier_periods is set if not gets right one
}
