% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{get_freq_adjustment}
\alias{get_freq_adjustment}
\title{Get change of frequency}
\usage{
get_freq_adjustment(date_type, frequency_number)
}
\arguments{
\item{date_type}{Date Type}

\item{frequency_number}{Original Frequency Number}
}
\value{
adjusted frequency
}
\description{
Get change of frequency
}
