% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_frequency_number}
\alias{get_frequency_number}
\title{Gets the right frequency numbers}
\usage{
get_frequency_number(date_type)
}
\arguments{
\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns frequency_number
}
\description{
Gets the right frequency numbers
}
