% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_data.R
\name{get_full_data_tbl}
\alias{get_full_data_tbl}
\title{Final function to create the full_data_tbl}
\usage{
get_full_data_tbl(
  data_tbl,
  combo_cleanup_date,
  combo_variables,
  clean_outliers,
  clean_missing_values,
  date_type,
  external_regressors,
  forecast_approach,
  frequency_number,
  forecast_horizon,
  hist_start_date,
  hist_end_date,
  pad_value,
  target_log_transformation,
  xregs_future_values_tbl
)
}
\arguments{
\item{data_tbl}{data table from previous pipe}

\item{combo_cleanup_date}{Date to clean up combo after}

\item{combo_variables}{List of combo variables}

\item{clean_outliers}{Is clean outliers out of data}

\item{clean_missing_values}{Is clean missing values}

\item{date_type}{date type from previous stage}

\item{external_regressors}{existing external_regressors}

\item{forecast_approach}{forecasting approach}

\item{frequency_number}{frequency number}

\item{forecast_horizon}{forecast horizon}

\item{hist_start_date}{historical start date}

\item{hist_end_date}{historical end date}

\item{pad_value}{Value for padding}

\item{target_log_transformation}{Log transformation}

\item{xregs_future_values_tbl}{external regressor table}
}
\value{
full_data_tbl
}
\description{
Final function to create the full_data_tbl
}
