% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_gluon_ts_frequency}
\alias{get_gluon_ts_frequency}
\title{Gets the right gluon ts frequency}
\usage{
get_gluon_ts_frequency(date_type)
}
\arguments{
\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns gluon_ts_frequency
}
\description{
Gets the right gluon ts frequency
}
