% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_kfold_tune_grid}
\alias{get_kfold_tune_grid}
\title{Get tuning grid k fold CV}
\usage{
get_kfold_tune_grid(train_data, wkflw, parallel = FALSE)
}
\arguments{
\item{train_data}{Training Data}

\item{wkflw}{Workflow Objet from previous stage}

\item{parallel}{Allow Parallal (Default False)}
}
\value{
gives the model fit
}
\description{
Get tuning grid k fold CV
}
