% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_lag_periods}
\alias{get_lag_periods}
\title{Gets the right lag periods}
\usage{
get_lag_periods(lag_periods, date_type, forecast_horizon)
}
\arguments{
\item{lag_periods}{lag_periods override}

\item{date_type}{year, quarter, month, week, day}

\item{forecast_horizon}{horion input from user}
}
\value{
Returns lag_periods
}
\description{
Checks if lag_periods is set if not gets right one
}
