% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_latin_hypercube_grid}
\alias{get_latin_hypercube_grid}
\title{Get grid_latin_hypercube}
\usage{
get_latin_hypercube_grid(model_spec)
}
\arguments{
\item{model_spec}{Model Spec Obj}
}
\value{
gives the latin hypercube grid
}
\description{
Get grid_latin_hypercube
}
