% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{get_model_functions}
\alias{get_model_functions}
\title{Gets list of pre-existing models}
\usage{
get_model_functions(models_to_run, model_not_to_run, run_deep_learning)
}
\arguments{
\item{models_to_run}{List of models to run}

\item{model_not_to_run}{List of models NOT to run}

\item{run_deep_learning}{Deep Learning Models}
}
\value{
uses models_to_run and models_not_to_run and returns correct list
}
\description{
Gets list of pre-existing models
}
