% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data.R
\name{get_modelling_ready_tbl}
\alias{get_modelling_ready_tbl}
\title{Gets the modelling ready table}
\usage{
get_modelling_ready_tbl(
  data_tbl,
  external_regressors,
  hist_end_date,
  combo_cleanup_date,
  combo_variables
)
}
\arguments{
\item{data_tbl}{standardized data table}

\item{external_regressors}{all external regressors}

\item{hist_end_date}{historical end date of data}

\item{combo_cleanup_date}{date to remove combos without value}

\item{combo_variables}{list of combo variables}
}
\value{
list of external regressors with future values
}
\description{
Gets the modelling ready table
}
