% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{get_not_all_data_models}
\alias{get_not_all_data_models}
\title{Get list of all-data models}
\usage{
get_not_all_data_models()
}
\value{
List of all data models
}
\description{
Get list of all-data models
}
