% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_data.R
\name{get_poly_trans_clean}
\alias{get_poly_trans_clean}
\title{Function to perform outlier cleaning and polynomial transformation}
\usage{
get_poly_trans_clean(
  df,
  clean_outliers,
  clean_missing_values,
  frequency_number,
  external_regressors
)
}
\arguments{
\item{df}{data frame}

\item{clean_outliers}{clean_outliers}

\item{clean_missing_values}{clean missing values}

\item{frequency_number}{frequency number}

\item{external_regressors}{external regressors}
}
\value{
cleaned_data
}
\description{
Function to perform outlier cleaning and polynomial transformation
}
