% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_info.R
\name{get_project_info}
\alias{get_project_info}
\title{Get project info}
\usage{
get_project_info(project_name = NULL, storage_object = NULL, path = NULL)
}
\arguments{
\item{project_name}{Name used to group similar runs under a
single project name.}

\item{storage_object}{Used to store outputs during a run to other
storage services in Azure. Could be a storage container object from
the 'AzureStor' package to connect to ADLS blob storage or a
OneDrive/SharePoint object from the 'Microsoft365R' package to connect
to a OneDrive folder or SharePoint site. Default of NULL will save outputs
to the local file system.}

\item{path}{String showing what file path the outputs should be written to.
Default of NULL will write the outputs to a temporary directory within R,
which will delete itself after the R session closes.}
}
\value{
Data frame of run log information
}
\description{
Lets you get all of the logging associated with a specific project or run.
}
\examples{
\donttest{
project_info <- set_project_info(
  project_name = "finn_forecast",
  combo_variables = c("Store", "Product"),
  target_variable = "Sales",
  date_type = "month"
)

project_info_tbl <- get_project_info(
  project_name = "finn_forecast"
)
}
}
