% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_recipie_combo}
\alias{get_recipie_combo}
\title{Gets a recipe with combo}
\usage{
get_recipie_combo(train_data)
}
\arguments{
\item{train_data}{Training Data}
}
\value{
combo recipie
}
\description{
Gets a recipe with combo
}
