% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_recipie_configurable}
\alias{get_recipie_configurable}
\title{Gets a recipe that adjusts based on parameters}
\usage{
get_recipie_configurable(
  train_data,
  fiscal_year_start,
  date_rm_regex_final,
  mutate_adj_half = TRUE,
  rm_date = "plain",
  step_nzv = "nzv",
  norm_date_adj_year = FALSE,
  dummy_one_hot = TRUE,
  character_factor = FALSE,
  center_scale = FALSE,
  one_hot = FALSE
)
}
\arguments{
\item{train_data}{Training Data}

\item{fiscal_year_start}{Start of Fiscal Year}

\item{date_rm_regex_final}{Date removal RegEx Final}

\item{mutate_adj_half}{parameter to add date adjustment}

\item{rm_date}{"plain", "with_adj", "with_adj_index"}

\item{step_nzv}{(Zero or NearZero Variance) "zv", "nzv", or none}

\item{norm_date_adj_year}{normalize date & year}

\item{dummy_one_hot}{one hot encoding}

\item{character_factor}{is character factor}

\item{center_scale}{Center and scale}

\item{one_hot}{True or False}
}
\value{
configurable recipie
}
\description{
Gets a recipe that adjusts based on parameters
}
