% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_recipie_fiscal_quarter_adj}
\alias{get_recipie_fiscal_quarter_adj}
\title{Gets a recipe that adjusts for fiscal year start}
\usage{
get_recipie_fiscal_quarter_adj(
  train_data,
  fiscal_year_start,
  date_rm_regex_final,
  model_type,
  one_hot = FALSE
)
}
\arguments{
\item{train_data}{Training Data}

\item{fiscal_year_start}{Start of Fiscal Year}

\item{date_rm_regex_final}{Date removal RegEx Final}

\item{model_type}{Model Type}
}
\value{
simple recipie
}
\description{
Gets a recipe that adjusts for fiscal year start
}
