% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_recipie_fiscal_year_adj}
\alias{get_recipie_fiscal_year_adj}
\title{Gets a recipe that adjusts for fiscal year start}
\usage{
get_recipie_fiscal_year_adj(train_data, fiscal_year_start, date_rm_regex_final)
}
\arguments{
\item{train_data}{Training Data}

\item{fiscal_year_start}{Start of Fiscal Year}

\item{date_rm_regex_final}{Date removal RegEx Final}
}
\value{
simple recipie
}
\description{
Gets a recipe that adjusts for fiscal year start
}
