% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_recipie_simple}
\alias{get_recipie_simple}
\title{Gets a simple recipe}
\usage{
get_recipie_simple(train_data)
}
\arguments{
\item{train_data}{Training Data}
}
\value{
simple recipie
}
\description{
Gets a simple recipe
}
