% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_resample_kfold}
\alias{get_resample_kfold}
\title{Get resample kFold CV}
\usage{
get_resample_kfold(train_data)
}
\arguments{
\item{train_data}{Training Data}
}
\value{
gives the resample kFold CV object
}
\description{
Get resample kFold CV
}
