% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_resample_tscv}
\alias{get_resample_tscv}
\title{Get resample Time Series CV}
\usage{
get_resample_tscv(train_data, tscv_initial, horizon, back_test_spacing)
}
\arguments{
\item{train_data}{Training Data}

\item{tscv_initial}{TS Cross Validation Initalization}

\item{horizon}{Horizon}

\item{back_test_spacing}{Back Testing Spacing}
}
\value{
gives the resample TS CV object
}
\description{
Get resample Time Series CV
}
