% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_resample_tune_grid}
\alias{get_resample_tune_grid}
\title{Get tuning grid with resample}
\usage{
get_resample_tune_grid(
  train_data,
  tscv_initial,
  horizon,
  back_test_spacing,
  wkflw,
  parallel = FALSE
)
}
\arguments{
\item{train_data}{Training Data}

\item{tscv_initial}{TS Cross Validation Initalization}

\item{horizon}{Horizon}

\item{back_test_spacing}{Back Testing Spacing}

\item{wkflw}{Workflow Objet from previous stage}

\item{parallel}{Allow Parallal (Default False)}
}
\value{
gives the model fit
}
\description{
Get tuning grid with resample
}
