% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_rolling_window_periods}
\alias{get_rolling_window_periods}
\title{Gets the right rolling window periods}
\usage{
get_rolling_window_periods(rolling_window_periods, date_type)
}
\arguments{
\item{rolling_window_periods}{rolling_window_periods override}

\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns rolling_window_periods
}
\description{
Checks if rolling_window_periods is set if not gets right one
}
