% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{get_seasonal_adjustment_models}
\alias{get_seasonal_adjustment_models}
\title{Get list of seasonal correction}
\usage{
get_seasonal_adjustment_models()
}
\value{
List of models that need seasonal adjustment
}
\description{
Get list of seasonal correction
}
