% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_seasonal_periods}
\alias{get_seasonal_periods}
\title{Gets the seasonal periods}
\usage{
get_seasonal_periods(date_type)
}
\arguments{
\item{date_type}{year, quarter, month, week, day}
}
\value{
Returns seasonal_periods
}
\description{
Gets the seasonal periods
}
