% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_simple_orchestration}
\alias{get_simple_orchestration}
\title{Get simple orchestration}
\usage{
get_simple_orchestration(train_data, frequency, fnToCall, parsnip_engine)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}

\item{fnToCall}{usually a modeltime function e.x. modeltime::arima_reg}

\item{parsnip_engine}{ex. 'auto_arima'}
}
\value{
gives the model fit
}
\description{
Get simple orchestration
}
