% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_read_write_data.R
\name{get_trained_models}
\alias{get_trained_models}
\title{Get Final Trained Models}
\usage{
get_trained_models(run_info)
}
\arguments{
\item{run_info}{run info}
}
\value{
table of final trained models
}
\description{
Get Final Trained Models
}
