% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_forecast_run.R
\name{get_transfer_functions}
\alias{get_transfer_functions}
\title{Fetches a list of parallel transfer functions}
\usage{
get_transfer_functions()
}
\value{
List of parallel transfer functions
}
\description{
Fetches a list of parallel transfer functions
}
