% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_tune_grid}
\alias{get_tune_grid}
\title{Get tuning grid with resample}
\usage{
get_tune_grid(train_data, wkflw, isBoost, resamples, parallel, isMetrics)
}
\arguments{
\item{train_data}{Training Data}

\item{wkflw}{Workflow}

\item{isBoost}{isBoost params}

\item{resamples}{Resamples}

\item{parallel}{Parallel}

\item{isMetrics}{is metrics}
}
\description{
Get tuning grid with resample
}
