% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_workflow_simple}
\alias{get_workflow_simple}
\title{Gets a simple workflow from model}
\usage{
get_workflow_simple(model_spec, recipe_spec)
}
\arguments{
\item{model_spec}{Model Spec A}

\item{recipe_spec}{year, quarter, month, week, day}
}
\value{
dplyr workflow spec
}
\description{
Gets a simple workflow from model
}
