% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data.R
\name{get_xreg_future_values_list}
\alias{get_xreg_future_values_list}
\title{Gets the list of external regressors with future values}
\usage{
get_xreg_future_values_list(data_tbl, external_regressors, hist_end_date)
}
\arguments{
\item{data_tbl}{standardized data table}

\item{external_regressors}{all external regressors}

\item{hist_end_date}{historical end date of data}
}
\value{
list of external regressors with future values
}
\description{
Gets the list of external regressors with future values
}
