% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_data.R
\name{get_xregs_future_values_tbl}
\alias{get_xregs_future_values_tbl}
\title{Function to create the xregs_future_values_tbl}
\usage{
get_xregs_future_values_tbl(xregs_future_values_tbl, forecast_approach)
}
\arguments{
\item{xregs_future_values_tbl}{existing external_regressors}

\item{forecast_approach}{forecasting approach}
}
\value{
xregs_future_values_tbl
}
\description{
Function to create the xregs_future_values_tbl
}
