% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{glmnet}
\alias{glmnet}
\title{GLM Net Function}
\usage{
glmnet(
  train_data,
  parallel,
  model_type = "single",
  horizon,
  tscv_initial,
  date_rm_regex,
  fiscal_year_start,
  back_test_spacing
)
}
\arguments{
\item{train_data}{Training Data}

\item{parallel}{Parallel Version or not}

\item{model_type}{"single" "ensemble" etc.}

\item{horizon}{Horizon of model}

\item{tscv_initial}{tscv initialization}

\item{date_rm_regex}{Date removal Regex}

\item{fiscal_year_start}{Fiscal Year Start}

\item{back_test_spacing}{Back Testing Spacing}
}
\value{
Get the GLM Net
}
\description{
GLM Net Function
}
\keyword{internal}
