% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_glmnet.R
\name{glmnet_multistep}
\alias{glmnet_multistep}
\title{GLMNET Multistep Horizon}
\usage{
glmnet_multistep(
  mode = "regression",
  mixture = NULL,
  penalty = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{mixture}{mixture}

\item{penalty}{penalty}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\value{
Get Multistep Horizon GLMNET model
}
\description{
GLMNET Multistep Horizon
}
\keyword{internal}
