% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_batch_parallel.R
\name{init_azure_parallel_within}
\alias{init_azure_parallel_within}
\title{Call back function within forecast for init}
\usage{
init_azure_parallel_within()
}
\description{
Sets the right configuration during init
}
