% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_parallel.R
\name{init_parallel_within}
\alias{init_parallel_within}
\title{Call back function within forecast for init}
\usage{
init_parallel_within(type)
}
\arguments{
\item{type}{Type of parallel processing being done}
}
\description{
Sets the right configuration during init
}
