% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{invoke_forecast_function}
\alias{invoke_forecast_function}
\title{Invoke model function with params}
\usage{
invoke_forecast_function(
  fn_to_invoke,
  train_data,
  frequency,
  horizon,
  parallel,
  seasonal_period,
  tscv_inital,
  date_rm_regex,
  back_test_spacing,
  fiscal_year_start,
  model_type
)
}
\arguments{
\item{fn_to_invoke}{Function to invoke}

\item{train_data}{Training Data}

\item{frequency}{Frequency used}

\item{horizon}{Horizon of forecast}

\item{parallel}{Is Include Parallel}

\item{seasonal_period}{Seasonal Peiod}

\item{tscv_inital}{Time Series CV Initialization}

\item{date_rm_regex}{Date Remove Regex}

\item{back_test_spacing}{Back Test Spacing}

\item{fiscal_year_start}{Fiscal Year Start}

\item{model_type}{Model Type}
}
\value{
Forecast Object to be used down stream
}
\description{
Invoke model function with params
}
