% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lightgbm}
\alias{lightgbm}
\title{Light GBM Model}
\usage{
lightgbm(train_data, fiscal_year_start, parallel)
}
\arguments{
\item{train_data}{Training Data}

\item{fiscal_year_start}{Fiscal Year Start}

\item{parallel}{should it be parallel}
}
\value{
Get the LightGBM based model
}
\description{
Light GBM Model
}
