% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_env_info.R
\name{load_env_info}
\alias{load_env_info}
\title{Load Environment Info}
\usage{
load_env_info(reticulate_environment)
}
\arguments{
\item{reticulate_environment}{Input python environment
ex. to connect to gluonts}
}
\description{
Checks if reticulate environment is needed
}
