% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mars}
\alias{mars}
\title{MARS Model}
\usage{
mars(
  train_data,
  parallel,
  model_type = "single",
  date_rm_regex,
  fiscal_year_start,
  pca
)
}
\arguments{
\item{train_data}{Training Data}

\item{parallel}{Parallel Version or not}

\item{model_type}{"single" "ensemble" etc.}

\item{date_rm_regex}{Date removal Regex}

\item{fiscal_year_start}{Fiscal Year Start}
}
\value{
Get the GLM Net
}
\description{
MARS Model
}
\keyword{internal}
