% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{mars_multistep}
\alias{mars_multistep}
\title{MARS Multistep Horizon}
\usage{
mars_multistep(
  mode = "regression",
  num_terms = NULL,
  prod_degree = NULL,
  prune_method = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{num_terms}{The number of features that will be retained in
the final model, including the intercept.}

\item{prod_degree}{The highest possible interaction degree.}

\item{prune_method}{The pruning method.}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\value{
Get Multistep Horizon MARS model
}
\description{
MARS Multistep Horizon
}
\keyword{internal}
