% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{mars_multistep_fit_impl}
\alias{mars_multistep_fit_impl}
\title{Bridge MARS Multistep Modeling function}
\usage{
mars_multistep_fit_impl(
  x,
  y,
  nprune = NULL,
  degree = 1L,
  pmethod = "backward",
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{nprune}{The number of features that will be retained in
the final model, including the intercept.}

\item{degree}{The highest possible interaction degree.}

\item{pmethod}{The pruning method.}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\description{
Bridge MARS Multistep Modeling function
}
\keyword{internal}
