% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{meanf}
\alias{meanf}
\title{Mean Forecast}
\usage{
meanf(train_data, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}
}
\value{
Get Mean Forecast Model
}
\description{
Mean Forecast
}
\keyword{internal}
