% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_train_models.R
\name{model_hyperparameters}
\alias{model_hyperparameters}
\title{Get model hyperparameters}
\usage{
model_hyperparameters(
  model_workflow_tbl,
  model_recipe_tbl,
  num_hyperparameters = 5
)
}
\arguments{
\item{model_workflow_tbl}{model workflow table}

\item{model_recipe_tbl}{model recipe table}

\item{num_hyperparameters}{number of hyperparameter combinations}
}
\value{
table of model hyperparameters
}
\description{
Get model hyperparameters
}
