% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_train_models.R
\name{model_workflows}
\alias{model_workflows}
\title{Gets model workflows}
\usage{
model_workflows(
  model_recipe_tbl,
  models_to_run = NULL,
  models_not_to_run = NULL,
  run_deep_learning = FALSE,
  pca = FALSE
)
}
\arguments{
\item{model_recipe_tbl}{}

\item{pca}{}
}
\value{
Returns back_test_scenarios and hist_periods_80
}
\description{
Gets model workflows
}
