% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nbeats}
\alias{nbeats}
\title{nbeats model}
\usage{
nbeats(train_data, horizon, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{frequency}{Frequency of Data}
}
\value{
Get nbeats Model
}
\description{
nbeats model
}
\keyword{internal}
