% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nnetar}
\alias{nnetar}
\title{nnetar model}
\usage{
nnetar(
  train_data,
  horizon,
  frequency,
  parallel,
  tscv_initial,
  back_test_spacing
)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{frequency}{Frequency of Data}

\item{parallel}{Parallel}

\item{tscv_initial}{TS CV Initialization}

\item{back_test_spacing}{Back Test Spacing}
}
\value{
Get nnetar Model
}
\description{
nnetar model
}
\keyword{internal}
