% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nnetar_xregs}
\alias{nnetar_xregs}
\title{nnetar model}
\usage{
nnetar_xregs(
  train_data,
  horizon,
  frequency,
  parallel,
  tscv_initial,
  date_rm_regex,
  fiscal_year_start,
  back_test_spacing,
  pca
)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{frequency}{Frequency of Data}

\item{parallel}{Parallel}

\item{tscv_initial}{TS CV Initialization}

\item{date_rm_regex}{Date RM Regex}

\item{fiscal_year_start}{Fiscal Year Start}

\item{back_test_spacing}{Back Test Spacing}
}
\value{
Get nnetar Model
}
\description{
nnetar model
}
\keyword{internal}
