% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_cubist.R
\name{predict.cubist_multistep_fit_impl}
\alias{predict.cubist_multistep_fit_impl}
\title{Predict custom cubist model}
\usage{
\method{predict}{cubist_multistep_fit_impl}(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}
}
\value{
predictions
}
\description{
Predict custom cubist model
}
\keyword{internal}
