% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_glmnet.R
\name{predict.glmnet_multistep_fit_impl}
\alias{predict.glmnet_multistep_fit_impl}
\title{Predict custom glmnet model}
\usage{
\method{predict}{glmnet_multistep_fit_impl}(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}
}
\value{
predictions
}
\description{
Predict custom glmnet model
}
\keyword{internal}
