% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{predict.mars_multistep_fit_impl}
\alias{predict.mars_multistep_fit_impl}
\title{Predict custom mars model}
\usage{
\method{predict}{mars_multistep_fit_impl}(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}
}
\value{
predictions
}
\description{
Predict custom mars model
}
\keyword{internal}
