% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_rbf.R
\name{predict.svm_rbf_multistep_fit_impl}
\alias{predict.svm_rbf_multistep_fit_impl}
\title{Predict custom svm_rbf model}
\usage{
\method{predict}{svm_rbf_multistep_fit_impl}(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}
}
\value{
predictions
}
\description{
Predict custom svm_rbf model
}
\keyword{internal}
