% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{predict.xgboost_multistep_fit_impl}
\alias{predict.xgboost_multistep_fit_impl}
\title{Predict custom xgboost model}
\usage{
\method{predict}{xgboost_multistep_fit_impl}(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}
}
\value{
predictions
}
\description{
Predict custom xgboost model
}
\keyword{internal}
