% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acr_scott_custom_model.R
\name{print.acr_scott_custom_model}
\alias{print.acr_scott_custom_model}
\title{Print the Scott custom model specification}
\usage{
\method{print}{acr_scott_custom_model}(x, ...)
}
\arguments{
\item{x}{A Scott custom model specification.}

\item{...}{Additional printer arguments.}
}
\value{
The specification invisibly, after printing its parameters.
}
\description{
Print the Scott custom model specification
}
