% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_fast_model.R
\name{print.arima_fast_model}
\alias{print.arima_fast_model}
\title{Print an adaptive fast ARIMA model specification}
\usage{
\method{print}{arima_fast_model}(x, ...)
}
\arguments{
\item{x}{An adaptive fast ARIMA model specification.}

\item{...}{Additional arguments.}
}
\value{
The model specification, invisibly.
}
\description{
Print an adaptive fast ARIMA model specification
}
\keyword{internal}
