% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos2_model.R
\name{print.chronos2_model}
\alias{print.chronos2_model}
\title{Print custom Chronos 2 model}
\usage{
\method{print}{chronos2_model}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{Additional arguments}
}
\value{
Prints model info
}
\description{
Print custom Chronos 2 model
}
\keyword{internal}
