% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos_bolt_tiny_model.R
\name{print.chronos_bolt_tiny_model}
\alias{print.chronos_bolt_tiny_model}
\title{Print custom Chronos Bolt Tiny model}
\usage{
\method{print}{chronos_bolt_tiny_model}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{Additional arguments}
}
\value{
Prints model info
}
\description{
Print custom Chronos Bolt Tiny model
}
\keyword{internal}
