% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_cubist.R
\name{print.cubist_multistep}
\alias{print.cubist_multistep}
\title{Print custom cubist model}
\usage{
\method{print}{cubist_multistep}(x, ...)
}
\value{
Prints model info
}
\description{
Print custom cubist model
}
\keyword{internal}
