% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_cubist.R
\name{print.cubist_multistep_fit_impl}
\alias{print.cubist_multistep_fit_impl}
\title{Print fitted custom cubist model}
\usage{
\method{print}{cubist_multistep_fit_impl}(x, ...)
}
\value{
prints custom model
}
\description{
Print fitted custom cubist model
}
\keyword{internal}
