% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_glmnet.R
\name{print.glmnet_multistep}
\alias{print.glmnet_multistep}
\title{Print custom glmnet model}
\usage{
\method{print}{glmnet_multistep}(x, ...)
}
\value{
Prints model info
}
\description{
Print custom glmnet model
}
\keyword{internal}
