% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_glmnet.R
\name{print.glmnet_multistep_fit_impl}
\alias{print.glmnet_multistep_fit_impl}
\title{Print fitted custom glmnet model}
\usage{
\method{print}{glmnet_multistep_fit_impl}(x, ...)
}
\value{
prints custom model
}
\description{
Print fitted custom glmnet model
}
\keyword{internal}
