% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{print.mars_multistep}
\alias{print.mars_multistep}
\title{Print custom mars model}
\usage{
\method{print}{mars_multistep}(x, ...)
}
\value{
Prints model info
}
\description{
Print custom mars model
}
\keyword{internal}
