% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{print.mars_multistep_fit_impl}
\alias{print.mars_multistep_fit_impl}
\title{Print fitted custom mars model}
\usage{
\method{print}{mars_multistep_fit_impl}(x, ...)
}
\value{
prints custom model
}
\description{
Print fitted custom mars model
}
\keyword{internal}
