% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_poly.R
\name{print.svm_poly_multistep}
\alias{print.svm_poly_multistep}
\title{Print custom svm_poly model}
\usage{
\method{print}{svm_poly_multistep}(x, ...)
}
\value{
Prints model info
}
\description{
Print custom svm_poly model
}
\keyword{internal}
