% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_rbf.R
\name{print.svm_rbf_multistep}
\alias{print.svm_rbf_multistep}
\title{Print custom svm_rbf model}
\usage{
\method{print}{svm_rbf_multistep}(x, ...)
}
\value{
Prints model info
}
\description{
Print custom svm_rbf model
}
\keyword{internal}
