% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_rbf.R
\name{print.svm_rbf_multistep_fit_impl}
\alias{print.svm_rbf_multistep_fit_impl}
\title{Print fitted custom svm_rbf model}
\usage{
\method{print}{svm_rbf_multistep_fit_impl}(x, ...)
}
\value{
prints custom model
}
\description{
Print fitted custom svm_rbf model
}
\keyword{internal}
