% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{print.timegpt}
\alias{print.timegpt}
\title{Print TimeGPT model specification}
\usage{
\method{print}{timegpt}(x, ...)
}
\arguments{
\item{x}{TimeGPT model specification}

\item{...}{Additional arguments}
}
\value{
Invisible model specification
}
\description{
Print TimeGPT model specification
}
