% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_model.R
\name{print.timegpt_model}
\alias{print.timegpt_model}
\title{Print custom TimeGPT model}
\usage{
\method{print}{timegpt_model}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{Additional arguments}
}
\value{
Prints model info
}
\description{
Print custom TimeGPT model
}
\keyword{internal}
