% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm_model.R
\name{print.timesfm_model}
\alias{print.timesfm_model}
\title{Print custom TimesFM model}
\usage{
\method{print}{timesfm_model}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{Additional arguments}
}
\value{
Prints model info
}
\description{
Print custom TimesFM model
}
\keyword{internal}
