% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm_model.R
\name{print.timesfm_model_fit_impl}
\alias{print.timesfm_model_fit_impl}
\title{Print custom TimesFM fit object}
\usage{
\method{print}{timesfm_model_fit_impl}(x, ...)
}
\arguments{
\item{x}{model fit object}

\item{...}{Additional arguments}
}
\value{
Prints fit info
}
\description{
Print custom TimesFM fit object
}
\keyword{internal}
