% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{print.xgboost_multistep}
\alias{print.xgboost_multistep}
\title{Print custom xgboost model}
\usage{
\method{print}{xgboost_multistep}(x, ...)
}
\value{
Prints model info
}
\description{
Print custom xgboost model
}
\keyword{internal}
