% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{print.xgboost_multistep_fit_impl}
\alias{print.xgboost_multistep_fit_impl}
\title{Print fitted custom xgboost model}
\usage{
\method{print}{xgboost_multistep_fit_impl}(x, ...)
}
\value{
prints custom model
}
\description{
Print fitted custom xgboost model
}
\keyword{internal}
