% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prophet}
\alias{prophet}
\title{prophet model}
\usage{
prophet(train_data, horizon, parallel, tscv_initial, back_test_spacing)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{parallel}{Parallel}

\item{tscv_initial}{TS CV Initalization}

\item{back_test_spacing}{Back Test Spacing}
}
\value{
Get prophet Model
}
\description{
prophet model
}
\keyword{internal}
