% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prophet_boost}
\alias{prophet_boost}
\title{prophet boost}
\usage{
prophet_boost(
  train_data,
  horizon,
  parallel,
  tscv_initial,
  date_rm_regex,
  fiscal_year_start,
  back_test_spacing,
  pca
)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{parallel}{Parallel}

\item{tscv_initial}{TS CV Initalization}

\item{date_rm_regex}{Date RM Regex}

\item{fiscal_year_start}{Fiscal Year Start}

\item{back_test_spacing}{Back Test Spacing}
}
\value{
Get prophet boost Model
}
\description{
prophet boost
}
\keyword{internal}
