% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prophet_xregs}
\alias{prophet_xregs}
\title{prophet xregs}
\usage{
prophet_xregs(
  train_data,
  horizon,
  parallel,
  tscv_initial,
  date_rm_regex,
  fiscal_year_start,
  back_test_spacing,
  pca
)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{parallel}{Parallel}

\item{tscv_initial}{TS CV Initialization}

\item{date_rm_regex}{Date RM Regex}

\item{fiscal_year_start}{Fiscal Year Start}

\item{back_test_spacing}{Back Test Spacing}
}
\value{
Get prophet xregs Model
}
\description{
prophet xregs
}
\keyword{internal}
