% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_run_info.R
\name{set_run_info}
\alias{set_run_info}
\title{Set up finnts submission}
\usage{
set_run_info(
  experiment_name = "finn_fcst",
  run_name = "finn_fcst",
  storage_object = NULL,
  path = NULL,
  data_output = "csv",
  object_output = "rds"
)
}
\arguments{
\item{experiment_name}{experiment name}

\item{run_name}{run name}

\item{storage_object}{storage object}

\item{path}{path}

\item{data_output}{data output}

\item{object_output}{object output}
}
\value{
A list of run information
}
\description{
Set up finnts submission
}
\examples{
\donttest{
run_info <- set_run_info(experiment_name = "finn forecast", 
                         run_name = "test_run", 
                         path = "test")
}
}
