% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{snaive}
\alias{snaive}
\title{SNaive}
\usage{
snaive(train_data, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}
}
\value{
Get SNaive Forecast Model
}
\description{
SNaive
}
