% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stlm_arima}
\alias{stlm_arima}
\title{STLM Arima}
\usage{
stlm_arima(train_data, seasonal_period)
}
\arguments{
\item{train_data}{Training Data}

\item{seasonal_period}{Seasonal Period}
}
\value{
Get STLM Arima Forecast Model
}
\description{
STLM Arima
}
\keyword{internal}
