% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stlm_ets}
\alias{stlm_ets}
\title{STLM ETS}
\usage{
stlm_ets(train_data, seasonal_period)
}
\arguments{
\item{train_data}{Training Data}

\item{seasonal_period}{Seasonal Period}
}
\value{
Get STLM ETS Forecast Model
}
\description{
STLM ETS
}
