% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_poly.R
\name{svm_poly_multistep}
\alias{svm_poly_multistep}
\title{SVM-POLY Multistep Horizon}
\usage{
svm_poly_multistep(
  mode = "regression",
  cost = NULL,
  degree = NULL,
  scale_factor = NULL,
  margin = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{cost}{A positive number for the cost of predicting
a sample within or on the wrong side of the margin.}

\item{degree}{A positive number for polynomial degree.}

\item{scale_factor}{A positive number for the polynomial
scaling factor.}

\item{margin}{A positive number for the epsilon in the SVM
insensitive loss function}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\value{
Get Multistep Horizon SVM-POLY model
}
\description{
SVM-POLY Multistep Horizon
}
\keyword{internal}
