% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_poly.R
\name{svm_poly_multistep_fit_impl}
\alias{svm_poly_multistep_fit_impl}
\title{Bridge SVM-POLY Multistep Modeling function}
\usage{
svm_poly_multistep_fit_impl(
  x,
  y,
  C = double(1),
  degree = integer(1),
  scale = double(1),
  epsilon = double(1),
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{C}{A positive number for the cost of predicting
a sample within or on the wrong side of the margin.}

\item{degree}{A positive number for polynomial degree.}

\item{scale}{A positive number for the polynomial
scaling factor.}

\item{epsilon}{A positive number for the epsilon in the SVM
insensitive loss function}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\description{
Bridge SVM-POLY Multistep Modeling function
}
\keyword{internal}
