% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_poly.R
\name{svm_poly_multistep_predict_impl}
\alias{svm_poly_multistep_predict_impl}
\title{Bridge prediction Function for SVM-POLY Multistep Horizon Models}
\usage{
svm_poly_multistep_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}

\item{...}{Additional \code{parsnip}-related options, depending on the
value of \code{type}. Arguments to the underlying model's prediction
function cannot be passed here (use the \code{opts} argument instead).
Possible arguments are:
\itemize{
\item \code{interval}: for \code{type} equal to \code{"survival"} or \code{"quantile"}, should
interval estimates be added, if available? Options are \code{"none"}
and \code{"confidence"}.
\item \code{level}: for \code{type} equal to \code{"conf_int"}, \code{"pred_int"}, or \code{"survival"},
this is the parameter for the tail area of the intervals
(e.g. confidence level for confidence intervals).
Default value is \code{0.95}.
\item \code{std_error}: for \code{type} equal to \code{"conf_int"} or \code{"pred_int"}, add
the standard error of fit or prediction (on the scale of the
linear predictors). Default value is \code{FALSE}.
\item \code{quantile}: for \code{type} equal to \code{quantile}, the quantiles of the
distribution. Default is \code{(1:9)/10}.
\item \code{eval_time}: for \code{type} equal to \code{"survival"} or \code{"hazard"}, the
time points at which the survival probability or hazard is estimated.
}}
}
\value{
predictions
}
\description{
Bridge prediction Function for SVM-POLY Multistep Horizon Models
}
\keyword{internal}
