% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tabnet}
\alias{tabnet}
\title{TabNet}
\usage{
tabnet(train_data, parallel, fiscal_year_start, date_rm_regex, pca)
}
\arguments{
\item{train_data}{Training Data}

\item{parallel}{Parallel}

\item{fiscal_year_start}{Fiscal Year Start}

\item{date_rm_regex}{Date RM Regex}
}
\value{
Get Tab Net
}
\description{
TabNet
}
\keyword{internal}
