% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tbats}
\alias{tbats}
\title{Tbats}
\usage{
tbats(train_data, seasonal_period)
}
\arguments{
\item{train_data}{Training Data}

\item{seasonal_period}{Seasonal Period}
}
\value{
Get TBats
}
\description{
Tbats
}
\keyword{internal}
