% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{theta}
\alias{theta}
\title{Theta Model}
\usage{
theta(train_data, frequency)
}
\arguments{
\item{train_data}{Training Data}

\item{frequency}{Frequency of Data}
}
\value{
Get the Theta based model
}
\description{
Theta Model
}
\keyword{internal}
