% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{timegpt}
\alias{timegpt}
\title{TimeGPT model specification}
\usage{
timegpt(
  mode = "regression",
  forecast_horizon = NULL,
  external_regressors = NULL,
  api_key = NULL
)
}
\arguments{
\item{mode}{Model mode (regression)}

\item{forecast_horizon}{Number of periods to forecast}

\item{external_regressors}{List of external regressors}

\item{api_key}{TimeGPT API key}
}
\value{
TimeGPT model specification
}
\description{
TimeGPT model specification
}
