% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{timegpt_fit_impl}
\alias{timegpt_fit_impl}
\title{Fit TimeGPT model}
\usage{
timegpt_fit_impl(
  x,
  y,
  forecast_horizon = NULL,
  external_regressors = NULL,
  api_key = NULL,
  ...
)
}
\arguments{
\item{x}{Training data}

\item{y}{Target variable}

\item{...}{Additional arguments}
}
\value{
Fitted TimeGPT model object
}
\description{
Fit TimeGPT model
}
\keyword{internal}
