% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_model.R
\name{timegpt_model_fit_impl}
\alias{timegpt_model_fit_impl}
\title{Fit TimeGPT model}
\usage{
timegpt_model_fit_impl(
  x,
  y,
  forecast_horizon = NULL,
  external_regressors = NULL
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{forecast_horizon}{forecast horizon}

\item{external_regressors}{external regressors}
}
\value{
Fitted TimeGPT model object
}
\description{
Fit TimeGPT model
}
\keyword{internal}
