% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_model.R
\name{timegpt_model_predict_impl}
\alias{timegpt_model_predict_impl}
\title{Bridge prediction Function for TimeGPT Models}
\usage{
timegpt_model_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}

\item{...}{Additional arguments}
}
\value{
predictions
}
\description{
Bridge prediction Function for TimeGPT Models
}
\keyword{internal}
