% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{timegpt_predict_impl}
\alias{timegpt_predict_impl}
\title{Predict with TimeGPT model}
\usage{
timegpt_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{Fitted TimeGPT model}

\item{new_data}{New data for prediction}

\item{...}{Additional arguments}
}
\value{
Predictions
}
\description{
Predict with TimeGPT model
}
\keyword{internal}
