% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm_model.R
\name{timesfm_model}
\alias{timesfm_model}
\title{TimesFM model specification}
\usage{
timesfm_model(mode = "regression", forecast_horizon = NULL, frequency = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency of data}
}
\value{
Get TimesFM model
}
\description{
TimesFM model specification
}
\keyword{internal}
