% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm_model.R
\name{timesfm_model_fit_impl}
\alias{timesfm_model_fit_impl}
\title{Fit TimesFM model}
\usage{
timesfm_model_fit_impl(x, y, forecast_horizon = NULL, frequency = NULL)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency of data}
}
\value{
Fitted TimesFM model object
}
\description{
Since TimesFM is a foundation model with no local training,
the fit step simply stores the training data and parameters
for use during prediction.
}
\keyword{internal}
