% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm_model.R
\name{timesfm_model_predict_impl}
\alias{timesfm_model_predict_impl}
\title{Bridge prediction function for TimesFM Models}
\usage{
timesfm_model_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}

\item{...}{Additional arguments}
}
\value{
numeric vector of predictions
}
\description{
Prepares the training data, calls the TimesFM API, and returns
the point predictions.
}
\keyword{internal}
