% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_model_utility.R
\name{train_test_split}
\alias{train_test_split}
\title{Gets the train test splits}
\usage{
train_test_split(
  run_info,
  back_test_scenarios = NULL,
  back_test_spacing = NULL
)
}
\arguments{
\item{run_info}{run info}

\item{back_test_scenarios}{back test scenarios}

\item{back_test_spacing}{back test spacing}
}
\value{
Returns table of train test splits
}
\description{
Gets the train test splits
}
\keyword{internal}
