% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_cubist.R
\name{translate.cubist_multistep}
\alias{translate.cubist_multistep}
\title{Translate custom cubist model}
\usage{
\method{translate}{cubist_multistep}(x, engine = x$engine, ...)
}
\value{
translated model
}
\description{
Translate custom cubist model
}
\keyword{internal}
