% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_glmnet.R
\name{translate.glmnet_multistep}
\alias{translate.glmnet_multistep}
\title{Translate custom glmnet model}
\usage{
\method{translate}{glmnet_multistep}(x, engine = x$engine, ...)
}
\value{
translated model
}
\description{
Translate custom glmnet model
}
\keyword{internal}
