% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{translate.mars_multistep}
\alias{translate.mars_multistep}
\title{Translate custom mars model}
\usage{
\method{translate}{mars_multistep}(x, engine = x$engine, ...)
}
\value{
translated model
}
\description{
Translate custom mars model
}
\keyword{internal}
