% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_poly.R
\name{translate.svm_poly_multistep}
\alias{translate.svm_poly_multistep}
\title{Translate custom svm_poly model}
\usage{
\method{translate}{svm_poly_multistep}(x, engine = x$engine, ...)
}
\value{
translated model
}
\description{
Translate custom svm_poly model
}
\keyword{internal}
