% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_rbf.R
\name{translate.svm_rbf_multistep}
\alias{translate.svm_rbf_multistep}
\title{Translate custom svm_rbf model}
\usage{
\method{translate}{svm_rbf_multistep}(x, engine = x$engine, ...)
}
\value{
translated model
}
\description{
Translate custom svm_rbf model
}
\keyword{internal}
