% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{translate.xgboost_multistep}
\alias{translate.xgboost_multistep}
\title{Translate custom xgboost model}
\usage{
\method{translate}{xgboost_multistep}(x, engine = x$engine, ...)
}
\value{
translated model
}
\description{
Translate custom xgboost model
}
\keyword{internal}
