% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_tune_models.R
\name{tune_models}
\alias{tune_models}
\title{Tune model hyperparameters}
\usage{
tune_models(
  model_recipe_tbl,
  model_workflow_tbl,
  model_hyperparameter_tbl,
  model_train_test_tbl,
  run_global_models,
  run_local_models,
  global_model_recipes,
  combo_variables,
  parallel_processing,
  num_cores,
  seed = 123,
  batch_size = 10000
)
}
\arguments{
\item{model_recipe_tbl}{model recipe table}

\item{model_workflow_tbl}{model workflow table}

\item{model_train_test_tbl}{model train test split table}

\item{run_global_models}{run global models}

\item{run_local_models}{run local models}

\item{global_model_recipes}{global model recipes}

\item{combo_variables}{combo variables}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}

\item{seed}{seed number}

\item{batch_size}{batch size}

\item{model_hyparameter_tbl}{model hyperparameter table}
}
\value{
table
}
\description{
Tune model hyperparameters
}
\keyword{internal}
