% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_fast_model.R
\name{update.arima_fast_model}
\alias{update.arima_fast_model}
\title{Update an adaptive fast ARIMA model specification}
\usage{
\method{update}{arima_fast_model}(
  object,
  parameters = NULL,
  forecast_horizon = NULL,
  frequency = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{An adaptive fast ARIMA model specification.}

\item{parameters}{A parameter object or \code{NULL}.}

\item{forecast_horizon}{Number of periods to forecast.}

\item{frequency}{Frequency of the input series.}

\item{fresh}{If \code{TRUE}, return a specification containing only new values.}

\item{...}{Additional arguments.}
}
\value{
An updated adaptive fast ARIMA model specification.
}
\description{
Update an adaptive fast ARIMA model specification
}
\keyword{internal}
