% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos_bolt_tiny_model.R
\name{update.chronos_bolt_tiny_model}
\alias{update.chronos_bolt_tiny_model}
\title{Update parameter in custom Chronos Bolt Tiny model}
\usage{
\method{update}{chronos_bolt_tiny_model}(
  object,
  parameters = NULL,
  forecast_horizon = NULL,
  frequency = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency}

\item{fresh}{fresh}

\item{...}{extra args}
}
\value{
Updated model
}
\description{
Update parameter in custom Chronos Bolt Tiny model
}
\keyword{internal}
