% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_cubist.R
\name{update.cubist_multistep}
\alias{update.cubist_multistep}
\title{Update parameter in custom cubist model}
\usage{
\method{update}{cubist_multistep}(
  object,
  parameters = NULL,
  committees = NULL,
  neighbors = NULL,
  max_rules = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{committees}{committees}

\item{neighbors}{neighbors}

\item{max_rules}{max rules}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}

\item{fresh}{fresh}

\item{...}{extra args passed to cubist}
}
\value{
Updated model
}
\description{
Update parameter in custom cubist model
}
\keyword{internal}
