% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_glmnet.R
\name{update.glmnet_multistep}
\alias{update.glmnet_multistep}
\title{Update parameter in custom glmnet model}
\usage{
\method{update}{glmnet_multistep}(
  object,
  parameters = NULL,
  mixture = NULL,
  penalty = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{mixture}{mixture}

\item{penalty}{penalty}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}

\item{fresh}{fresh}

\item{...}{extra args passed to glmnet}
}
\value{
Updated model
}
\description{
Update parameter in custom glmnet model
}
\keyword{internal}
