% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_mars.R
\name{update.mars_multistep}
\alias{update.mars_multistep}
\title{Update parameter in custom mars model}
\usage{
\method{update}{mars_multistep}(
  object,
  parameters = NULL,
  num_terms = NULL,
  prod_degree = NULL,
  prune_method = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{num_terms}{The number of features that will be retained in
the final model, including the intercept.}

\item{prod_degree}{The highest possible interaction degree.}

\item{prune_method}{The pruning method.}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}

\item{fresh}{fresh}

\item{...}{extra args passed to mars}
}
\value{
Updated model
}
\description{
Update parameter in custom mars model
}
\keyword{internal}
