% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_svm_rbf.R
\name{update.svm_rbf_multistep}
\alias{update.svm_rbf_multistep}
\title{Update parameter in custom svm_rbf model}
\usage{
\method{update}{svm_rbf_multistep}(
  object,
  parameters = NULL,
  cost = NULL,
  rbf_sigma = NULL,
  margin = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{cost}{A positive number for the cost of predicting
a sample within or on the wrong side of the margin.}

\item{rbf_sigma}{A positive number for radial basis function.}

\item{margin}{A positive number for the epsilon in the SVM
insensitive loss function.}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}

\item{fresh}{fresh}

\item{...}{extra args passed to svm_rbf}
}
\value{
Updated model
}
\description{
Update parameter in custom svm_rbf model
}
\keyword{internal}
