% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{update.timegpt}
\alias{update.timegpt}
\title{Update parameter in custom TimeGPT model}
\usage{
\method{update}{timegpt}(
  object,
  parameters = NULL,
  forecast_horizon = NULL,
  external_regressors = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{forecast_horizon}{forecast horizon}

\item{external_regressors}{external regressors}

\item{fresh}{fresh}

\item{...}{extra args passed to TimeGPT}
}
\value{
Updated model
}
\description{
Update parameter in custom TimeGPT model
}
\keyword{internal}
