% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_model.R
\name{update.timegpt_model}
\alias{update.timegpt_model}
\title{Update parameter in custom TimeGPT model}
\usage{
\method{update}{timegpt_model}(
  object,
  parameters = NULL,
  forecast_horizon = NULL,
  fresh = FALSE,
  frequency = NULL,
  finetune_steps = NULL,
  finetune_depth = NULL,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{forecast_horizon}{forecast horizon}

\item{fresh}{fresh}

\item{frequency}{frequency}

\item{finetune_steps}{finetune steps}

\item{finetune_depth}{finetune depth}

\item{...}{extra args passed to TimeGPT}
}
\value{
Updated model
}
\description{
Update parameter in custom TimeGPT model
}
\keyword{internal}
