% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm_model.R
\name{update.timesfm_model}
\alias{update.timesfm_model}
\title{Update parameter in custom TimesFM model}
\usage{
\method{update}{timesfm_model}(
  object,
  parameters = NULL,
  forecast_horizon = NULL,
  frequency = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{forecast_horizon}{forecast horizon}

\item{frequency}{frequency}

\item{fresh}{fresh}

\item{...}{extra args}
}
\value{
Updated model
}
\description{
Update parameter in custom TimesFM model
}
\keyword{internal}
