% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{update.xgboost_multistep}
\alias{update.xgboost_multistep}
\title{Update parameter in custom xgboost model}
\usage{
\method{update}{xgboost_multistep}(
  object,
  parameters = NULL,
  mtry = NULL,
  trees = NULL,
  min_n = NULL,
  tree_depth = NULL,
  learn_rate = NULL,
  loss_reduction = NULL,
  sample_size = NULL,
  stop_iter = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{parameters}{parameters}

\item{mtry}{mtry}

\item{trees}{trees}

\item{min_n}{min_n}

\item{tree_depth}{tree depth}

\item{learn_rate}{learn rate}

\item{loss_reduction}{loss reduction}

\item{sample_size}{number for the number (or proportion) of data that is exposed to the fitting routine.}

\item{stop_iter}{The number of iterations without improvement before stopping}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}

\item{fresh}{fresh}

\item{...}{extra args passed to xgboost}
}
\value{
Updated model
}
\description{
Update parameter in custom xgboost model
}
\keyword{internal}
