% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_forecasting_inputs.R
\name{validate_forecasting_inputs}
\alias{validate_forecasting_inputs}
\title{Validate Forecasting Inputs}
\usage{
validate_forecasting_inputs(
  input_data,
  combo_variables,
  target_variable,
  external_regressors,
  forecast_horizon,
  date_type,
  hist_start_date,
  hist_end_date,
  combo_cleanup_date,
  fiscal_year_start,
  clean_missing_values,
  clean_outliers,
  back_test_scenarios,
  back_test_spacing,
  modeling_approach,
  forecast_approach,
  parallel_processing,
  run_model_parallel,
  azure_batch_credentials,
  max_model_average
)
}
\arguments{
\item{input_data}{Input Dataframe or Tibble}

\item{combo_variables}{Column headers forming a tuple}

\item{target_variable}{Column header to forecast}

\item{external_regressors}{Column header used as regressors}

\item{forecast_horizon}{Number of periods to forecast}

\item{date_type}{Type of the time variable}

\item{hist_start_date}{Input data start, default is earliest}

\item{hist_end_date}{Input data end, default is latest}

\item{combo_cleanup_date}{Date after which to trim zero series}

\item{fiscal_year_start}{Month number of start of fiscal year}

\item{clean_missing_values}{Cleaning missing values}

\item{clean_outliers}{Cleaning outliers in data}

\item{back_test_scenarios}{Auto or 1,2,3, etc.}

\item{back_test_spacing}{Auto or 1,2,3, etc.}

\item{modeling_approach}{Currently only accuracy is supported}

\item{forecast_approach}{Bottoms_up, grouped_hierarchy, standard_hierarchy}

\item{parallel_processing}{azure_batch, local_machine, none}

\item{run_model_parallel}{run hyperparameter search and model in parallel}

\item{azure_batch_credentials}{Azure Batch Credentials}

\item{max_model_average}{Maximum number of models to average}
}
\value{
Returns hist_start_date and hist_end_date
}
\description{
Validates if all the forecasting inputs are as expected
}
\details{
Function will stop execution and send an error otherwise
}
