% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{xgboost}
\alias{xgboost}
\title{XGBoost}
\usage{
xgboost(
  train_data,
  horizon,
  parallel,
  model_type = "single",
  tscv_initial,
  date_rm_regex,
  fiscal_year_start,
  back_test_spacing
)
}
\arguments{
\item{train_data}{Training Data}

\item{horizon}{Horizon}

\item{parallel}{Parallel}

\item{model_type}{Type of Model}

\item{tscv_initial}{TS CV Initalization}

\item{date_rm_regex}{Date RM Regex}

\item{fiscal_year_start}{Fiscal Year start}

\item{back_test_spacing}{Back Test Spacing}
}
\value{
Get XGBoost
}
\description{
XGBoost
}
