% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{xgboost_multistep}
\alias{xgboost_multistep}
\title{XGBOOST Multistep Horizon}
\usage{
xgboost_multistep(
  mode = "regression",
  mtry = NULL,
  trees = NULL,
  min_n = NULL,
  tree_depth = NULL,
  learn_rate = NULL,
  loss_reduction = NULL,
  sample_size = NULL,
  stop_iter = NULL,
  lag_periods = NULL,
  external_regressors = NULL,
  forecast_horizon = NULL,
  selected_features = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{mtry}{mtry}

\item{trees}{trees}

\item{min_n}{min_n}

\item{tree_depth}{tree depth}

\item{learn_rate}{learn rate}

\item{loss_reduction}{loss reduction}

\item{sample_size}{number for the number (or proportion) of data that is exposed to the fitting routine.}

\item{stop_iter}{The number of iterations without improvement before stopping}

\item{lag_periods}{lag periods}

\item{external_regressors}{external regressors}

\item{forecast_horizon}{forecast horizon}

\item{selected_features}{selected features}
}
\value{
Get Multistep Horizon XGBoost model
}
\description{
XGBOOST Multistep Horizon
}
\keyword{internal}
