% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistep_xgboost.R
\name{xgboost_multistep_predict_impl}
\alias{xgboost_multistep_predict_impl}
\title{Bridge prediction Function for XGBOOST Multistep Horizon Models}
\usage{
xgboost_multistep_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{model object}

\item{new_data}{input data to predict}

\item{...}{Additional arguments passed to \code{predict.xgb.Booster()}}
}
\value{
predictions
}
\description{
Bridge prediction Function for XGBOOST Multistep Horizon Models
}
\keyword{internal}
