% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_ensemble_models.R
\name{ensemble_models}
\alias{ensemble_models}
\title{Ensemble Models}
\usage{
ensemble_models(
  tuning_tbl,
  refit_tbl,
  model_train_test_tbl,
  date_type,
  num_hyperparameters = 5,
  parallel_processing = NULL,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{tuning_tbl}{hyperparameter tuning predictions}

\item{refit_tbl}{individual model predictions}

\item{model_train_test_tbl}{model train test split table}

\item{date_type}{date type}

\item{num_hyperparameters}{number of hyperparameters}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}

\item{seed}{seed number}
}
\value{
list with ensemble predictions and fitted models
}
\description{
Ensemble Models
}
\keyword{internal}
