% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_train_models.R
\name{ensemble_models}
\alias{ensemble_models}
\title{Ensemble Models}
\usage{
ensemble_models(
  tuning_tbl,
  refit_tbl,
  date_type,
  parallel_processing = NULL,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{seed}{seed number}

\item{model_recipe_tbl}{model recipe table}

\item{model_fit_tbl}{model fit table}

\item{model_train_test_tbl}{model train test split table}

\item{combo_variables}{combo variables}
}
\value{
table
}
\description{
Ensemble Models
}
\keyword{internal}
