% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_ensemble_models.R
\name{ensemble_models}
\alias{ensemble_models}
\title{Ensemble Models}
\usage{
ensemble_models(
  run_info,
  parallel_processing = NULL,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{run_info}{run info}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}

\item{seed}{seed number}
}
\value{
outputs are written to disk
}
\description{
Ensemble Models
}
\keyword{internal}
