% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_final_models.R
\name{final_models}
\alias{final_models}
\title{Select Best Models and Prep Final Outputs}
\usage{
final_models(
  refit_predictions_tbl,
  ensemble_predictions_tbl = NULL,
  average_predictions_tbl = NULL,
  model_train_test_tbl,
  parallel_processing = NULL,
  num_cores = NULL
)
}
\arguments{
\item{refit_predictions_tbl}{individual model predictions}

\item{ensemble_predictions_tbl}{ensemble model predictions}

\item{average_predictions_tbl}{model average predictions}

\item{model_train_test_tbl}{train test splits}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}
}
\value{
tbl with best model flag
}
\description{
Select Best Models and Prep Final Outputs
}
\keyword{internal}
