% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_final_models.R
\name{final_models}
\alias{final_models}
\title{Select Best Models and Prep Final Outputs}
\usage{
final_models(
  run_info,
  average_models = TRUE,
  max_model_average = 3,
  parallel_processing = NULL,
  num_cores = NULL
)
}
\arguments{
\item{run_info}{run info}

\item{average_models}{average models}

\item{max_model_average}{max model average}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}
}
\value{

}
\description{
Select Best Models and Prep Final Outputs
}
\keyword{internal}
