% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_time_series.R
\name{forecast_time_series}
\alias{forecast_time_series}
\title{Finn Forecast Framework}
\usage{
forecast_time_series(
  input_data,
  combo_variables,
  target_variable,
  date_type,
  forecast_horizon,
  external_regressors = NULL,
  run_name = "time_series_forecast",
  hist_start_date = NULL,
  hist_end_date = NULL,
  combo_cleanup_date = NULL,
  fiscal_year_start = 1,
  clean_missing_values = TRUE,
  clean_outliers = FALSE,
  back_test_scenarios = "auto",
  back_test_spacing = "auto",
  modeling_approach = "accuracy",
  forecast_approach = "bottoms_up",
  parallel_processing = "none",
  run_model_parallel = TRUE,
  azure_batch_credentials = NULL,
  azure_batch_cluster_config = NULL,
  azure_batch_cluster_delete = FALSE,
  target_log_transformation = FALSE,
  negative_fcst = FALSE,
  fourier_periods = NULL,
  lag_periods = NULL,
  rolling_window_periods = NULL,
  reticulate_environment = NULL,
  models_to_run = NULL,
  models_not_to_run = NULL,
  run_deep_learning = FALSE,
  run_all_data = TRUE,
  average_models = TRUE,
  max_model_average = 4,
  weekly_to_daily = TRUE
)
}
\arguments{
\item{input_data}{A data frame or tibble of historical time series data. Can also include external regressors for both
historical and future data.}

\item{combo_variables}{List of column headers within input data to be used to separate individual time series.}

\item{target_variable}{The column header formatted as a character value within input data you want to forecast.}

\item{date_type}{The date granularity of the input data. Finn accepts the following as a character string
day, week, month, quarter, year.}

\item{forecast_horizon}{Number of periods to forecast into the future.}

\item{external_regressors}{List of column headers within input data to be used as features in multivariate models.}

\item{run_name}{Name used when submitting jobs to external compute like Azure Batch. Formatted as a character string.}

\item{hist_start_date}{Date value of when your input_data starts. Default of NULL is to use earliest date value in
input_data.}

\item{hist_end_date}{Date value of when your input_data ends.Default of NULL is to use the latest date value in
input_data.}

\item{combo_cleanup_date}{Date value to remove individual time series that don't contain non-zero values after
that specified date. Default of NULL is to not remove any time series and attempt to forecast all of them.}

\item{fiscal_year_start}{Month number of start of fiscal year of input data, aids in building out date features.
Formatted as a numeric value. Default of 1 assumes fiscal year starts in January.}

\item{clean_missing_values}{Should missing values be inputted? Only inputes values for missing data within an
existing series, and does not add new values onto the beginning or end, but does provide a value of 0 for said
values.}

\item{clean_outliers}{Should outliers be cleaned and inputted with values more in line with historical data?}

\item{back_test_scenarios}{Number of specific back test folds to run when determining the best model.
Default of 'auto' will automatially choose the number of back tests to run based on historical data size,
which tries to always use a minimum of 80\% of the data when training a model.}

\item{back_test_spacing}{Number of periods to move back for each back test scenario. Default of 'auto'}

\item{modeling_approach}{test}

\item{forecast_approach}{test}

\item{parallel_processing}{test}

\item{run_model_parallel}{test}

\item{azure_batch_credentials}{test}

\item{azure_batch_cluster_config}{test}

\item{azure_batch_cluster_delete}{test}

\item{target_log_transformation}{test}

\item{negative_fcst}{test}

\item{fourier_periods}{test}

\item{lag_periods}{test}

\item{rolling_window_periods}{test}

\item{reticulate_environment}{test}

\item{models_to_run}{test}

\item{models_not_to_run}{test}

\item{run_deep_learning}{test}

\item{run_all_data}{test}

\item{average_models}{test}

\item{max_model_average}{test}

\item{weekly_to_daily}{test}
}
\description{
Finn Forecast Framework
}
