% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_iterate_forecast.R
\name{get_agent_forecast}
\alias{get_agent_forecast}
\title{Get the final best forecast for an agent}
\usage{
get_agent_forecast(agent_info)
}
\arguments{
\item{agent_info}{Agent info from \code{set_agent_info()}}
}
\value{
A tibble containing the final forecast for the agent.
}
\description{
This function retrieves the final forecast for a Finn agent after the forecast iteration process is complete.
}
\examples{
\dontrun{
# load example data
hist_data <- timetk::m4_monthly \%>\%
  dplyr::filter(date >= "2013-01-01") \%>\%
  dplyr::rename(Date = date) \%>\%
  dplyr::mutate(id = as.character(id))

# set up Finn project
project <- set_project_info(
  project_name = "Demo_Project",
  combo_variables = c("id"),
  target_variable = "value",
  date_type = "month"
)

# set up LLM
llm <- ellmer::chat_azure_openai(model = "gpt-4o-mini")

# set up agent info
agent_info <- set_agent_info(
  project_info = project,
  llm = llm,
  input_data = hist_data,
  forecast_horizon = 6
)

# run the forecast iteration process
iterate_forecast(
  agent_info = agent_info,
  max_iter = 3,
  weighted_mape_goal = 0.03
)

# get the final forecast for the agent
final_forecast <- get_agent_forecast(agent_info = agent_info)
}
}
