% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_iterate_forecast.R
\name{get_best_agent_run}
\alias{get_best_agent_run}
\title{Get the best run for an agent}
\usage{
get_best_agent_run(agent_info)
}
\arguments{
\item{agent_info}{A list containing agent information including project info and run ID.}
}
\value{
A tibble containing the best run information for the agent.
}
\description{
Get the best run for an agent
}
