% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_fit_simple}
\alias{get_fit_simple}
\title{Gets a simple fit}
\usage{
get_fit_simple(wflw_spec, train_data)
}
\arguments{
\item{wflw_spec}{Worflow Spec}

\item{train_data}{Training Data}
}
\value{
simple recipie
}
\description{
Gets a simple fit
}
