% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_read_write_data.R
\name{get_forecast_data}
\alias{get_forecast_data}
\title{Get Final Forecast Data}
\usage{
get_forecast_data(run_info, return_type = "df")
}
\arguments{
\item{run_info}{run info}

\item{return_type}{return type}
}
\value{
table of final forecast results
}
\description{
Get Final Forecast Data
}
