% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hierarchy_data.R
\name{get_log_transformation}
\alias{get_log_transformation}
\title{Function to perform log transformation}
\usage{
get_log_transformation(df, target_log_transformation)

get_log_transformation(df, target_log_transformation)
}
\arguments{
\item{df}{data frame}

\item{target_log_transformation}{variable to indicate log transformation}

\item{clean_outliers}{clean_outliers}

\item{clean_missing_values}{clean missing values}

\item{frequency_number}{frequency number}
}
\value{
full_data_tbl with or without log transformation

full_data_tbl with or without log transformation
}
\description{
Function to perform log transformation

Function to perform outlier cleaning
}
