% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_run_info.R
\name{get_run_info}
\alias{get_run_info}
\title{Get run info}
\usage{
get_run_info(
  experiment_name = "finn_fcst",
  run_name = NULL,
  storage_object = NULL,
  path = NULL,
  return_type = "df"
)
}
\arguments{
\item{experiment_name}{experiment name}

\item{run_name}{run name}

\item{storage_object}{storage object}

\item{path}{path}

\item{return_type}{return type}
}
\value{
table of run information
}
\description{
Get run info
}
\examples{
\donttest{
run_info <- set_run_info(experiment_name = "finn_forecast", 
                         run_name = "test_run", 
                         path = NULL)
                         
run_info_tbl <- get_run_info(experiment_name = "finn_forecast", 
                             run_name = "test_run")
}
}
