% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_iterate_forecast.R
\name{iterate_forecast}
\alias{iterate_forecast}
\title{Run the Finn Agent Forecast Iteration Process}
\usage{
iterate_forecast(
  agent_info,
  max_iter = 3,
  weighted_mape_goal = 0.03,
  parallel_processing = NULL,
  inner_parallel = FALSE,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{agent_info}{A list containing the agent information, including project info and LLMs.}

\item{max_iter}{Maximum number of iterations for the global model optimization.}

\item{weighted_mape_goal}{The goal for the weighted MAPE metric.}

\item{parallel_processing}{Logical indicating whether to use parallel processing.}

\item{inner_parallel}{Logical indicating whether to use inner parallel processing.}

\item{num_cores}{Number of cores to use for parallel processing.}

\item{seed}{Random seed for reproducibility.}
}
\description{
Run the Finn Agent Forecast Iteration Process
}
