% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_model_utility.R
\name{model_hyperparameters}
\alias{model_hyperparameters}
\title{Get model hyperparameters}
\usage{
model_hyperparameters(run_info, num_hyperparameters = 10)
}
\arguments{
\item{run_info}{run info}

\item{num_hyperparameters}{number of hyperparameter combinations}
}
\value{
table of model hyperparameters
}
\description{
Get model hyperparameters
}
\keyword{internal}
