% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_model_utility.R
\name{model_workflows}
\alias{model_workflows}
\title{Gets model workflows}
\usage{
model_workflows(
  model_recipe_tbl,
  models_to_run = NULL,
  models_not_to_run = NULL,
  run_deep_learning = FALSE,
  pca = FALSE
)
}
\arguments{
\item{model_recipe_tbl}{model recipe table}

\item{models_to_run}{models to run}

\item{models_not_to_run}{models not to run}

\item{run_deep_learning}{run deep learning models}

\item{pca}{pca}
}
\value{
Returns table of model workflows
}
\description{
Gets model workflows
}
\keyword{internal}
