% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_model_utility.R
\name{prep_models}
\alias{prep_models}
\title{Prep Models}
\usage{
prep_models(
  run_info,
  back_test_scenarios = NULL,
  back_test_spacing = NULL,
  models_to_run = NULL,
  models_not_to_run = NULL,
  run_deep_learning = FALSE,
  pca = FALSE,
  num_hyperparameters = 10
)
}
\arguments{
\item{run_info}{run info}

\item{back_test_scenarios}{back test scenarios}

\item{back_test_spacing}{back test spacing}

\item{models_to_run}{models to run}

\item{models_not_to_run}{models not to run}

\item{run_deep_learning}{run deep learning}

\item{pca}{pca}

\item{num_hyperparameters}{num hyperparameter}
}
\value{

}
\description{
Preps various aspects of run before training models. Things like train/test splits, creating hyperparameters, etc.
}
