% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{print.timegpt}
\alias{print.timegpt}
\title{Print custom TimeGPT model}
\usage{
\method{print}{timegpt}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{Additional arguments}
}
\value{
Prints model info
}
\description{
Print custom TimeGPT model
}
\keyword{internal}
