% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_refit_models.R
\name{refit_models}
\alias{refit_models}
\title{Refit Models}
\usage{
refit_models(
  run_info,
  combo_variables,
  parallel_processing = NULL,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{run_info}{run info}

\item{combo_variables}{combo variables}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}

\item{seed}{seed number}
}
\value{
list of individual model predictions and fitted models
}
\description{
Refit Models
}
\keyword{internal}
