% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_refit_models.R
\name{refit_models}
\alias{refit_models}
\title{Refit Models}
\usage{
refit_models(
  model_fit_tbl,
  model_recipe_tbl,
  model_train_test_tbl = NULL,
  combo_variables,
  parallel_processing = NULL,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{model_fit_tbl}{model fit table}

\item{model_recipe_tbl}{model recipe table}

\item{model_train_test_tbl}{model train test split table}

\item{combo_variables}{combo variables}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}

\item{seed}{seed number}
}
\value{
list of individual model predictions and fitted models
}
\description{
Refit Models
}
\keyword{internal}
