% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt.R
\name{timegpt}
\alias{timegpt}
\title{TimeGPT model specification}
\usage{
timegpt(
  mode = "regression",
  forecast_horizon = NULL,
  external_regressors = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{forecast_horizon}{forecast horizon}

\item{external_regressors}{external regressors}
}
\value{
Get TimeGPT model
}
\description{
TimeGPT model specification
}
\keyword{internal}
