% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_train_models.R
\name{train_models}
\alias{train_models}
\title{Train Individual Models}
\usage{
train_models(
  run_info,
  run_global_models,
  run_local_models,
  global_model_recipes,
  combo_variables,
  parallel_processing,
  num_cores,
  seed = 123
)
}
\arguments{
\item{run_info}{run info}

\item{run_global_models}{run global models}

\item{run_local_models}{run local models}

\item{global_model_recipes}{global model recipes}

\item{combo_variables}{combo variables}

\item{parallel_processing}{parallel processing}

\item{num_cores}{number of cores}

\item{seed}{seed number}
}
\value{
table
}
\description{
Train Individual Models
}
\keyword{internal}
