% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnts_0.2.0_train_models.R
\name{train_test_split}
\alias{train_test_split}
\title{Gets the train test splits}
\usage{
train_test_split(
  input_tbl,
  hist_end_date,
  date_type,
  forecast_horizon,
  back_test_scenarios = NULL,
  back_test_spacing = NULL
)
}
\arguments{
\item{input_tbl}{full data table}

\item{hist_end_date}{historical end date}

\item{forecast_horizon}{forecast horizon}

\item{back_test_scenarios}{back test scenarios}

\item{back_test_spacing}{back test spacing}
}
\value{
Returns back_test_scenarios and hist_periods_80
}
\description{
Gets the train test splits
}
\keyword{internal}
