% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_update_forecast.R
\name{update_forecast}
\alias{update_forecast}
\title{Update Forecast Agent}
\usage{
update_forecast(
  agent_info,
  weighted_mape_goal = 0.1,
  allow_iterate_forecast = FALSE,
  max_iter = 3,
  parallel_processing = NULL,
  inner_parallel = FALSE,
  num_cores = NULL,
  seed = 123
)
}
\arguments{
\item{agent_info}{A list containing the agent information.}

\item{weighted_mape_goal}{Numeric indicating the goal for the weighted MAPE.}

\item{allow_iterate_forecast}{Logical indicating if the forecast iteration should be allowed.}

\item{max_iter}{Numeric indicating the maximum number of iterations if iterate_forecast is ran.}

\item{parallel_processing}{Logical indicating if parallel processing should be used.}

\item{inner_parallel}{Logical indicating if inner parallel processing should be used.}

\item{num_cores}{Numeric indicating the number of cores to use for parallel processing.}

\item{seed}{Numeric seed for reproducibility.}
}
\value{
Nothing
}
\description{
This function updates the forecast agent with the latest data and inputs.
}
